requirements.txt
fastapi
uvicorn
google-genai
supabase
python-dotenv
python-multipart
